Fixed a bunch of UI stuff
Still need to retrieve additional information for the User on the view
profile page. Only have access to E-Mail and Username. Currently trying
to get a count of the number of crumbs they have to display. Also, found
a bug on the View My Crumbs page. Pressing back will exit the app.