Logout now closes the activity it is used in
When logging out using the action bar menu logout option, you could
press back on the login screen and it would go back to the activity you
pressed it in (Activities now finish correctly).  I believe the since
the id was the same in all of the xmls for the logout button, it would
logout using the wrong logout (i.e. if you logged out in the rankings
activity, it would use the logout from a different activity or
something).  Renaming the ids in the xml and onoptionitemselected for
each activity fixed it.